Remove percentage from pie chart
