+++ app/javascript/charts/information_access_by_gender_chart.js
@@ -1,38 +1,5 @@
 import { sum } from '../utils/array'
-
-let mapLabel = (chart, label, i) => {
-  let meta = chart.getDatasetMeta(0);
-  let ds = chart.data.datasets[0];
-  let arc = meta.data[i];
-  let custom = arc && arc.custom || {};
-  let getValueAtIndexOrDefault = Chart.helpers.getValueAtIndexOrDefault;
-  let arcOpts = chart.options.elements.arc;
-  let fill = custom.backgroundColor ? custom.backgroundColor : getValueAtIndexOrDefault(ds.backgroundColor, i, arcOpts.backgroundColor);
-  let stroke = custom.borderColor ? custom.borderColor : getValueAtIndexOrDefault(ds.borderColor, i, arcOpts.borderColor);
-  let bw = custom.borderWidth ? custom.borderWidth : getValueAtIndexOrDefault(ds.borderWidth, i, arcOpts.borderWidth);
-  let perc = parseFloat(ds.data[i] / chart.data.total*100).toFixed(2)
-
-  return {
-    text: `${label} (${perc}%)`,
-    fillStyle: fill,
-    strokeStyle: stroke,
-    lineWidth: bw,
-    hidden: isNaN(ds.data[i]) || meta.data[i].hidden,
-    index: i
-  }
-}
-
-let generateLabels = function(chart) {
-  let data = chart.data
-  return _.map(data.labels, mapLabel.bind(null, chart))
-}
-
-let isDisplay = function(context) {
-  let dataset = context.dataset;
-  let count = dataset.data.length;
-  let value = dataset.data[context.dataIndex];
-  return value > count * 1.5;
-}
+import { generateLabels, isDisplay } from '../data/piechart/defaults'
 
 export const genderInfo = () => {
   let type = 'pie',
